+++ doc/Changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.9.1] - UNRELEASED
+
+### Added
+
+* Embeddings extraction is now more flexible in order to support different models. For now, the parameters allow to configure the name of the tensor output, and to select one of two modes: `Raw` or `Token` (see crate documentation).
+* Added two examples for embeddings using `gte-multilingual-base`, that leverage these new options.
+
+
 ## [0.9.0] - 2025-02-01
 
 Initial release.